refactor [component] : 불필요한 파일 삭제 및 파일 정리
- 모달 관련 파일들을 Modal 폴더에 정리
- ModalContainer 의 하위 컴포넌트에 대한 파일과 ModalContainer를 ModalContainer 폴더에서 정리
- 불필요한 파일 삭제: CenterModalContext, ToastMOdalContext 삭제
- 컴포넌트에 대한 주석 추가
