+++ components/src/lib/components/Modal/index.tsx
@@ -1,0 +1,3 @@
+export { default as BottomModal } from './BottomModal';
+export { default as CenterModal } from './CenterModal';
+export { default as ToastModal } from './ToastModal';
