--- components/src/lib/components/ModalContainer.tsx
@@ -1,55 +1,0 @@
-import { MouseEvent } from 'react';
-import styled from 'styled-components';
-
-import '../styles/reset.css';
-import { ModalContainerContext } from '../contexts';
-import { useModalContext } from '../hooks';
-import { ModalButtonProps, ModalContainerProps } from '../types/modal';
-
-import Backdrop from './Backdrop';
-import Contents from './Contents';
-import ModalPortal from './ModalPortal';
-
-const ModalWrapper = styled.div`
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-`;
-
-function ModalContainer(props: ModalContainerProps) {
-  const { openModal, children, ...rest } = props;
-
-  return (
-    <>
-      {openModal && (
-        <ModalPortal>
-          <ModalContainerContext.Provider value={{ ...rest }}>
-            <ModalWrapper>{children}</ModalWrapper>
-          </ModalContainerContext.Provider>
-        </ModalPortal>
-      )}
-    </>
-  );
-}
-
-function ModalButton({ isCloseModal, children, onClick, ...rest }: ModalButtonProps) {
-  const { closeModal } = useModalContext(ModalContainerContext);
-
-  const handleClick = (e: MouseEvent<HTMLButtonElement>) => {
-    if (onClick) onClick(e);
-    if (isCloseModal) {
-      closeModal();
-    }
-  };
-
-  return (
-    <button {...rest} onClick={handleClick}>
-      {children}
-    </button>
-  );
-}
-ModalContainer.backdrop = Backdrop;
-ModalContainer.button = ModalButton;
-ModalContainer.contents = Contents;
-
-export default ModalContainer;
