+++ components/src/lib/components/ModalContainer/Backdrop.tsx
@@ -1,9 +1,9 @@
 import { MouseEvent, useEffect } from 'react';
 import styled from 'styled-components';
 
-import { BASIC_BACKGROUND_COLOR } from '../constants/modal';
-import { ModalContainerContext } from '../contexts';
-import { useModalContext } from '../hooks';
+import { BASIC_BACKGROUND_COLOR } from '../../constants/modal';
+import { ModalContainerContext } from '../../contexts';
+import { useModalContext } from '../../hooks';
 
 const ModalBackdrop = styled.div<{ $backgroundColor: string | undefined }>`
   position: fixed;
