+++ components/src/lib/components/ModalContainer/Contents.tsx
@@ -1,10 +1,10 @@
 import { ReactNode } from 'react';
 import styled from 'styled-components';
 
-import { BASIC_BACKGROUND_COLOR, BASIC_BORDER_RADIUS, BASIC_PADDING } from '../constants/modal';
-import { ModalContainerContext } from '../contexts';
-import { useModalContext } from '../hooks';
-import { Background, ModalPosition } from '../types/modal';
+import { BASIC_BACKGROUND_COLOR, BASIC_BORDER_RADIUS, BASIC_PADDING } from '../../constants/modal';
+import { ModalContainerContext } from '../../contexts';
+import { useModalContext } from '../../hooks';
+import { Background, ModalPosition } from '../../types/modal';
 
 export interface ModalContentsStyleProps {
   $borderRadius: string | undefined;
@@ -14,7 +14,10 @@ export interface ModalContentsStyleProps {
   $timeout?: number;
   $position?: ModalPosition;
 }
-
+/**
+ * Modal 내의 children을 감싸는 Contents의 동일한 스타일을 적용할 수 있도록 하는 스타일 컴포넌트,
+ * ModalContents를 상속받아서 모달 타입별로 다른 Contents를 만들 수 있다.
+ */
 export const ModalContents = styled.div<ModalContentsStyleProps>`
   -webkit-box-shadow: 0px 0px 18px 6px rgba(0, 0, 0, 0.19);
   box-shadow: 0px 0px 18px 6px rgba(0, 0, 0, 0.19);
@@ -31,7 +34,10 @@ export const ModalContents = styled.div<ModalContentsStyleProps>`
     min-width: 80vw;
   }
 `;
-
+/**
+ * Modal 내의 children을 감싸는 Contents의 동일한 스타일을 적용할 수 있도록 하는 컴포넌트
+ * @returns
+ */
 export default function Contents({ children }: { children: ReactNode }) {
   const {
     borderRadius = BASIC_BORDER_RADIUS,
