+++ components/src/lib/components/index.tsx
@@ -1,4 +1,6 @@
-export { default as BottomModal } from './BottomModal';
-export { default as CenterModal } from './CenterModal';
-export { default as ToastModal } from './ToastModal';
+import { BottomModal, CenterModal, ToastModal } from './Modal';
+
+export { default as CloseButtonIcon } from './CloseButtonIcon';
 export { default as ModalContainer } from './ModalContainer';
+export { default as ModalPortal } from './ModalPortal';
+export { BottomModal, CenterModal, ToastModal };
