--- components/src/lib/contexts/CenterModalContext.ts
@@ -1,9 +1,0 @@
-import { createContext } from 'react';
-
-export interface CenterModalContextType {
-  closeModal: () => void;
-}
-
-const CenterModalContext = createContext<CenterModalContextType | null>(null);
-
-export default CenterModalContext;
