--- components/src/lib/contexts/ToastModalContext.ts
@@ -1,8 +1,0 @@
-import { createContext } from 'react';
-
-import { ToastModalProps } from '../types/modal';
-
-interface TostModalContextType extends Omit<ToastModalProps, 'children' | 'openModal' | 'setOpenModal' | 'children'> {}
-const ToastModalContext = createContext<TostModalContextType | null>(null);
-
-export default ToastModalContext;
