+++ components/src/lib/contexts/index.ts
@@ -1,4 +1,2 @@
 export { default as BottomModalContext } from './BottomModalContext';
 export { default as ModalContainerContext } from './ModalContainerContext';
-export { default as CenterModalContext } from './CenterModalContext';
-export { default as ToastModalContext } from './ToastModalContext';
